Move feedback form to dedicated page, simplify popup to link
Creates /feedback page with the full form (Turnstile, message, email)
and replaces the popup form with a lightweight prompt linking to that
page. Re-enables the popup (ENABLED=true). This avoids loading
Turnstile in a popup where it caused blank white screen issues.
