+++ merger-tracker/frontend/src/App.jsx
@@ -19,6 +19,7 @@ import Digest from './pages/Digest';
 import NickTwort from './pages/NickTwort';
 import Analysis from './pages/Analysis';
 import PrivacyPolicy from './pages/PrivacyPolicy';
+import Feedback from './pages/Feedback';
 import NotFound from './pages/NotFound';
 
 function AppContent() {
@@ -44,6 +45,7 @@ function AppContent() {
             <Route path="/analysis" element={<Analysis />} />
             <Route path="/nick-twort" element={<NickTwort />} />
             <Route path="/privacy" element={<PrivacyPolicy />} />
+            <Route path="/feedback" element={<Feedback />} />
             <Route path="*" element={<NotFound />} />
           </Routes>
         </main>
